+++ app/src/main/java/com/rummenigged/popcorntime/view/seriesList/model/SeriesUiState.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.view.seriesList
+package com.rummenigged.popcorntime.view.seriesList.model
 
 data class SeriesUiState(
     val isLoading: Boolean = false,
